frontend/requirements.txt
chainlit>=2.9.6
httpx>=0.28.1
python-dotenv>=1.0.0